+++ src/components/i18n/es.ts
@@ -1,12 +1,12 @@
 import { I18n } from '../../ol-pdf-printer';
 
 export const es: I18n = {
-    printPdf: 'Imprimir PDF',
+    printPdf: 'Exportar PDF',
     pleaseWait: 'Por favor espere...',
     almostThere: 'Ya casi...',
     error: 'Error al generar pdf',
     errorImage: 'Ocurrió un error al tratar de cargar una imagen',
-    printing: 'Imprimiendo',
+    printing: 'Exportando',
     cancel: 'Cancelar',
     close: 'Cerrar',
     print: 'Exportar',